Add useDrawer hook updating when localstorage is updated not through the hook
Source: https://github.com/msupply-foundation/openmsupply-client/commit/03c4e7e8a8d0c91638be490401b77ea0339f9a66

+++ client/packages/common/src/localStorage/LocalStorage.ts
@@ -40,6 +40,11 @@ class LocalStorage {
     key: StorageKey,
     value: LocalStorageRecord[StorageKey]
   ): void {
+    const existingValue = this.getItem(key);
+
+    // Short circuit exit if values are unchanged
+    if (existingValue === value) return;
+
     const stringified = JSON.stringify(value);
     localStorage.setItem(this.createStorageKey(key), stringified);
